demux: fix es offset issue [1/1]
PD#SWPL-51051

Problem:
 es offset issue

Solution:
 es offset not alignm with es start

Verify:
 none

+++ drivers/amlogic/dvb/demux/sc2_demux/mem_desc.h
@@ -37,7 +37,7 @@ struct chan_id {
 	unsigned int memdescs_phy;
 	unsigned int r_offset;
 	unsigned long memdescs_map;
-	unsigned long last_w_addr;
+	unsigned int last_w_addr;
 	unsigned int tee_handle;
 
 	/*just for DVR sec direct mem*/
